+++ app/api/driver_api.py
@@ -2,7 +2,7 @@
 
 from fastapi import APIRouter
 
-from app.core.db_driver_enum import DBTypesEnum
+from app.core.enum.db_driver import DBTypesEnum
 from app.core.exceptions import APIException
 from app.core.status import CommonCode
 from app.schemas.driver_info import DriverInfo
